Deemphasize 'Cloud' usage in the Sync page (#255)
I'm trying to deemphasize the term 'cloud' here because I think it
implies that it's going somewhere unknown. It should be clear that it's
going to a users installed server, and nowhere else. Pikapods and
fly.io, I think would be classic examples where where the term 'cloud'
makes sense but wouldn't make sense if someone had it installed on a
local PC, NAS or Raspberry Pi.

I left a lot of the document alone further down because it seems like
there may be error messages within the app that use the term 'cloud'.

Maybe add an additional explainer in the section that explains the cloud
file?

**Came up in a discord support issue.