Clean up the homepage feature list (#376)
some of the features on the homepage seemed there just to pad out the
list. The list if decently long and interesting so I removed Payees and
Split transactions. Those seem pretty expected and not that interesting
as a called out feature.

I modified the budgeting feature to be more specific. Just saying we
have budgeting is a bit redundant for a budgeting app.