Add a download page to the website (#359)
This is a super rough start for a download page for the website. Please
make contributions since this is pretty bad so far. Can you tell Im not
a front end dev? :laughing:

In order for the links to always work we either need to:
* Modify our electron build to not add versions to the filename
* Add some way for the page to know the current version and use that in
the requested filename

For now they are hard coded with the version so they wont work next
month unless something changes.

For now the OS logos are colored so that they can be seen in both light
and dark modes. Im sure there is a better way than making them purple.