Upating Local install docs for people on a clean clone (#645)
Issue reported here: https://github.com/actualbudget/actual/issues/4377

https://deploy-preview-645.www.actualbudget.org/docs/install/local

It turns out that yarn doesn't let you run scripts without first
installing dependencies, so users will have to type it in the install
manually.

The other alternative is to run the script with npm but that would add
confusion.

I think yarn workspace focus is the best option to pick but here was the
options:

- Instruct users to ```yarn install``` (which would install all deps) -
1.07GB
- Instruct users to ```yarn workspaces focus @actual-app/sync-server
--production``` (which is a long command) - 60.6MB
- Create a new shell script to run the command ```
./install-server.sh```

In future we might want to consider packaging the server up into an
executable. That way the user wont have to install node/deal with any of
this stuff.

Here's an example packager - i've never used it:
https://github.com/nexe/nexe